refactor. Database 삭제 및 repository계층으로 기능이전
